Kokoro: use upstream kokoro package via pip instead of vendoring model source
Addresses PR review feedback to include the model from upstream rather than
copy-pasting the whole thing. Removes the four vendored source files under
kokoro/pytorch/src/ (~1000 lines) and instead installs the upstream `kokoro`
PyPI package (pinned 0.9.4, --no-deps to skip the misaki/espeak g2p stack that
only the unused KPipeline needs).

The small set of TT-specific edits that cannot live in an unmodified upstream
checkout is applied as runtime monkey-patches in tt_patches.py (fixed per-token
duration to dodge the tt-mlir round_nearest_even gap, deterministic phase/
additive noise for PCC, LSTM pack/pad bypass, dtype pins, and the classic
weight_norm needed to match the released checkpoint keys).

- requirements.nodeps.txt: kokoro==0.9.4 (installed --no-deps)
- requirements.txt: trigger file (all of kokoro's import-time deps are already
  guaranteed by the tt-xla env; loguru is a pjrt-plugin-tt dependency)
- tt_patches.py imports KModel while bypassing kokoro/__init__ (misaki), and
  resolves the pip package via distribution metadata so it is not shadowed by
  the same-named tt-forge-models kokoro/ model directory on sys.path.

Verified on n150: model compiles and runs end-to-end; test XFAILs on the
genuine bf16 waveform-PCC wall (pcc=0.138, unchanged from before the refactor).
